wip - refactor
