+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsmanageadjudicationsapi/services/reported/DamagesService.kt
@@ -1,0 +1,40 @@
+package uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.reported
+
+import org.springframework.stereotype.Service
+import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.controllers.DamageRequestItem
+import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.dtos.ReportedAdjudicationDto
+import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.entities.ReportedDamage
+import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.repositories.ReportedAdjudicationRepository
+import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.security.AuthenticationFacade
+import uk.gov.justice.digital.hmpps.hmppsmanageadjudicationsapi.services.OffenceCodeLookupService
+import javax.transaction.Transactional
+
+@Transactional
+@Service
+class DamagesService(
+  reportedAdjudicationRepository: ReportedAdjudicationRepository,
+  offenceCodeLookupService: OffenceCodeLookupService,
+  authenticationFacade: AuthenticationFacade,
+) : ReportedAdjudicationBaseService(
+  reportedAdjudicationRepository, offenceCodeLookupService, authenticationFacade
+) {
+  fun updateDamages(adjudicationNumber: Long, damages: List<DamageRequestItem>): ReportedAdjudicationDto {
+    val reportedAdjudication = findByAdjudicationNumber(adjudicationNumber)
+    val reporter = authenticationFacade.currentUsername!!
+    val toPreserve = reportedAdjudication.damages.filter { it.reporter != reporter }
+
+    reportedAdjudication.damages.clear()
+    reportedAdjudication.damages.addAll(toPreserve)
+    reportedAdjudication.damages.addAll(
+      damages.filter { it.reporter == reporter }.map {
+        ReportedDamage(
+          code = it.code,
+          details = it.details,
+          reporter = reporter
+        )
+      }
+    )
+
+    return save(reportedAdjudication)
+  }
+}
